refactor: remove dependency on iroha_wasm_builder from iroha cli and client
This moves the `construct_executor` utility function from `iroha_cli` to `test_network`, as it only gets used in tests
