+++ client/Cargo.toml
@@ -106,10 +106,6 @@ once_cell = { workspace = true }
 
 trybuild = { workspace = true }
 
-[build-dependencies]
-eyre = { workspace = true }
-iroha_wasm_builder = { workspace = true }
-
 [[bench]]
 name = "torii"
 harness = false
