+++ client/examples/register_1000_triggers.rs
@@ -11,10 +11,10 @@ use iroha::{
 use iroha_data_model::parameter::{Parameter, SmartContractParameter};
 use iroha_genesis::{GenesisBlock, GenesisBuilder};
 use iroha_primitives::unique_vec;
-use irohad::samples::{construct_executor, get_config};
+use irohad::samples::get_config;
 use test_network::{
-    get_chain_id, get_key_pair, wait_for_genesis_committed_with_max_retries, Peer as TestPeer,
-    PeerBuilder, TestClient, TestRuntime,
+    construct_executor, get_chain_id, get_key_pair, wait_for_genesis_committed_with_max_retries,
+    Peer as TestPeer, PeerBuilder, TestClient, TestRuntime,
 };
 use test_samples::gen_account_in;
 use tokio::runtime::Runtime;
