+++ core/test_network/Cargo.toml
@@ -19,6 +19,7 @@ iroha_data_model.workspace = true
 iroha_primitives.workspace = true
 iroha_logger.workspace = true
 iroha_genesis.workspace = true
+iroha_wasm_builder.workspace = true
 test_samples.workspace = true
 
 eyre.workspace = true
